Feature scrum 145 reponer cartas (#17)
* SCRUM-104 buscar_figuras_en_tablero : agrupar fichas por color y dividir en subconjuntos adyacentes + endpoint para probar funcionalidad

* SCRUM-104-buscar_figuras_en_tablero verificacion primeras 10 figuras

* SCRUM104 buscar_figuras_en_tablero : added is_fig11 ... is_fig25

* SCRUM104 buscar_figuras_en_tablero, arreglo error unhashable-typeset

* SCRUM104 buscaer_figuras_tablero broadcast data_figuras encontradas, (anda medio mal)

* SCRUM104 buscar_figuras_en_tablero : figuras se buscan luego de iniciar partida, solo se buscan las figuras cuyas cartas de figura son visibles en el juego, correccion is_fig1

* SCRUM 104 buscar figuras en el tablero: Ahora se computa buscar figuras cuando se ejecuta iniciar y abandonar partida, mensaje de web socket corrgido, error en figura 18 corregido

* SCRUM104 mensaje socket mandado correctamente

* SCRUM104 identificador idFig en mensaje de websocket

* SCRUM104 buscar figuras en tablero: mensaje socket con posiciones = (fila,columna), en lugar de (columna, fila)

* SCRUM-96 verificación completar figura válida

* SCRUM-96 actualizar figuras en el tablero al jugar movimiento y declarar figuras

* SCRUM-96 agrego test

* SCRUM-96 cambio respuesta al completar una figura

* SCRUM-96 cambio parametros del body request

* SCRUM-94 cambio parametros body request nuevamente

* SCRUM-96 los status_code que se devuelven

* SCRUM-96 actualizar figuras al deshacer movimientos y declarar mal una figura

* SCRUM-96 cambio respuesta por websockets

* SCRUM-145-reponer-cartas modificación de websockets game

* SCRUM-145-reponer-cartas se añadió 1 esquema de websockets

* SCRUM-145-reponer-cartas metodos para reponer cartas

* SCRUM-145-reponer-cartas modificacion de logica en endpoint pasar turno

* SCRUM-145-reponer-cartas modificaciones service abandonar_partida

* SCRUM-145-reponer-cartas eliminación de sleep innecesario

* SCRUM-145-reponer-cartas agregar esquema de WebSocket

* SCRUM-145-reponer-cartas modificación de orden en el que entrego los turnos al iniciar partida

* SCRUM-145-reponer-cartas metodos para reponer cartas de movimientos al jugador

* SCRUM-145-reponer-cartas eliminar movimientos parciales en el momento que un jugador juega una figura formada en el tablero

* SCRUM-145-reponer-cartas modificación del orden de los turnos

* SCRUM-145-reponer-carta modificacion de turno

* SCRUM-145-reponer-cartas pruebas

* SCRUM-145-reponer-cartas prueba

* SCRUM-145-reponer-cartas esquemas de WebSocket

* SCRUM-145-reponer-cartas eliminacion de carta de figura cuando se encuentra en el tablero y se utiliza

* SCRUM-145-reponer-cartas modificacion reposicion cartas de figuras

* SCRUM-145-reponer-cartas eliminacion  comentario

* SCRUM-145-reponer-cartas eliminacion de comentarios innecesarios

* SCRUM-145-reponer-cartas eliminacion de archivos

* SCRUM-145-reponer-cartas eliminación de comentarios

* SCRUM-145-reponer-cartas eliminacion de comentarios innecesarios

* SCRUM-145-reponer-cartas arreglo figura 6

---------

+++ app/services/partida_service.py
@@ -21,7 +21,6 @@ class PartidaService:
         return partidas 
     
     def obtener_partida(self, id_partida: int, db: Session):        
-        print(id_partida)
         partida = db.query(Partida).filter(Partida.id == id_partida).first()
         if partida is None:
             raise HTTPException(status_code=404, detail=f"No existe ninguna partida con id {id_partida}, error 1")
@@ -146,6 +145,7 @@ class PartidaService:
         cartas_figuras = obtener_cartas_figuras(id_partida, db)
         fichas = fichas_service.obtener_fichas(id_partida, db)
         orden = obtener_id_jugadores(id_partida, db)
+        
         response = {
             "type": "IniciarPartida",
             "fichas": fichas, 
@@ -156,7 +156,15 @@ class PartidaService:
         return response
     
     def pasar_turno(self, id_partida: int, id_jugador, db: Session):  
-        partida = self.obtener_partida(id_partida,db)  
+        
+        partida = self.obtener_partida(id_partida,db) 
+        if partida is None:
+            print("la partida no existe")
+            
+        jugador = db.query(Jugador).filter(Jugador.id == id_jugador).first()
+        if jugador is None :
+            print(f"el jugador con id {id_jugador} no pertenece a la partida")
+                
         tablero = partida.tablero
         turno_actual = id_jugador
 
@@ -181,25 +189,58 @@ class PartidaService:
             if not jugador:
                 raise HTTPException(status_code=404, detail=f"No existe ningún jugador con id {id_jugador}")
 
-            jugador_partida = db.query(Jugador_Partida).filter_by(id_partida=id_partida, id_jugador=id_jugador).first()
-            if not jugador_partida:
-                raise HTTPException(status_code=404, detail=f"El jugador {id_jugador} no está en la partida {id_partida}")
+            jugador_partida = db.query(Jugador_Partida).filter(Jugador_Partida.id_partida == id_partida).all()
 
             cantidad_jugadores = obtener_cantidad_jugadores(id_partida, db)
 
             if partida.activa:
                 if cantidad_jugadores == 2:
-                    await self.eliminar_partida(id_partida, db)
+                    
+                    if (jugador_partida):
+                        for jp in jugador_partida:
+                            db.delete(jp)
+                    db.commit()        
+                    
+                    tablero = db.query(Tablero).filter(Tablero.id_partida == id_partida).first()
+
+                    if tablero is None:
+                        raise HTTPException(status_code=404, detail=f"abandonar No existe tablero asociado a id : {id_partida}")
+
+                    db.query(Ficha).filter(Ficha.id_tablero == tablero.id).delete()
+
+                    # Eliminar los tableros  y las cartas asociadas a la partida
+                    db.query(Tablero).filter(Tablero.id_partida == id_partida).delete()
+                    db.query(CartasFigura).filter(CartasFigura.id_partida == id_partida).delete()
+                    db.query(CartaMovimientos).filter(CartaMovimientos.id_partida == id_partida).delete()
+                    
+                    jugadores = self.obtener_jugadores(id_partida, db)
+                    for jugador in jugadores:
+                        db.delete(jugador)
+                    
+                    db.delete(partida)
+                    
                 else:
-                    db.delete(jugador_partida)
-                    db.commit()
+                    db.query(Jugador_Partida).filter(Jugador_Partida.id_jugador == id_jugador,
+                                                     Jugador_Partida.id_partida == id_partida).delete()
+                    
+                    tablero = db.query(Tablero).filter(Tablero.id_partida == id_partida).first()
+
+                    if tablero is None:
+                        raise HTTPException(status_code=404, detail=f"abandonar No existe tablero asociado a id : {id_partida}")
+
+                    db.query(CartasFigura).filter(CartasFigura.id_partida == id_partida,
+                                                  CartasFigura.id_jugador == id_jugador).delete()
+                    db.query(CartaMovimientos).filter(CartaMovimientos.id_partida == id_partida,
+                                                      CartaMovimientos.id_jugador == id_jugador).delete()
+                    db.delete(jugador)
+
             else:
                 if cantidad_jugadores > 1 and id_jugador != partida.id_owner:
+                    db.delete(jugador)
                     db.delete(jugador_partida)
                 else:
                     await self.eliminar_partida(id_partida, db)  # Elimina si el owner abandona antes de comenzar
 
-            jugador.jugando = False
             db.commit()
 
         except SQLAlchemyError as e:
@@ -217,16 +258,17 @@ class PartidaService:
 
     async def eliminar_partida(self, id_partida: int, db: Session):
         try:
-            # Obtener la partida
             partida = db.query(Partida).filter(Partida.id == id_partida).first()
 
             if not partida:
                 raise HTTPException(status_code=404, detail="Partida no encontrada")
-
-            # Eliminar las fichas asociadas a los tableros de la partida
-            tableros = db.query(Tablero).filter(Tablero.id_partida == id_partida).all()
-            for tablero in tableros:
-                db.query(Ficha).filter(Ficha.id_tablero == tablero.id).delete()
+            
+            tablero = db.query(Tablero).filter(Tablero.id_partida == id_partida).first()
+            
+            if tablero is None:
+                raise HTTPException(status_code=404, detail=f"No existe tablero asociado a id : {id_partida}")
+            
+            db.query(Ficha).filter(Ficha.id_tablero == tablero.id).delete()
 
             # Eliminar los tableros  y las cartas asociadas a la partida
             db.query(Tablero).filter(Tablero.id_partida == id_partida).delete()
